Converted server to MVC
