+++ server/src/main.rs
@@ -5,114 +5,37 @@ extern crate rocket;
 #[macro_use]
 extern crate serde_derive;
 
-use bson::{bson, doc};
-use mongodb::options::ClientOptions;
+mod models {
+    pub mod game_model;
+    pub mod user_model;
+}
+
+mod controllers {
+    pub mod game_controller;
+    pub mod user_controller;
+}
+
+use controllers::{game_controller, user_controller};
+use mongodb::{options::ClientOptions, sync::Client, sync::Database};
 use rocket::http::Method::{Get, Post};
-use rocket::http::RawStr;
-use rocket_contrib::json::Json;
 use rocket_cors::{AllowedOrigins, CorsOptions};
-mod database;
-#[derive(FromForm, Debug, Serialize, Deserialize)]
-pub struct UserInfo {
-    pub username: String,
-    pub password: String,
+
+pub struct MyMongo {
+    db: Database,
 }
 
-#[derive(FromForm, Debug, Serialize, Deserialize)]
-pub struct ScoreUpdate {
-    pub username: String,
-    pub game: u8, // 0 connect 4, 1 toot
-    pub win: u8,  // 0 loss, 1 win, 2 tie
-}
+impl MyMongo {
+    pub fn new() -> Result<MyMongo, mongodb::error::Error> {
+        let mut client_options = ClientOptions::parse("mongodb://localhost:27017")?;
 
-#[derive(FromForm, Debug, Serialize, Deserialize)]
-pub struct GameInfo {
-    pub username: String,
-    pub xo_wins: i32,
-    pub xo_loss: i32,
-    pub xo_ties: i32,
-    pub to_wins: i32,
-    pub to_loss: i32,
-    pub to_ties: i32,
-}
+        // Manually set an option.
+        client_options.app_name = Some("Server".to_string());
 
-#[get("/")]
-fn index() -> &'static str {
-    "Hello, world!"
-}
+        // Get a handle to the deployment.
+        let conn = Client::with_options(client_options)?;
+        let db = conn.database("421ServerDB");
 
-#[post("/new_user", format = "application/json", data = "<user>")]
-fn new_user(user: Json<UserInfo>) -> Json<String> {
-    match database::MyMongo::new() {
-        Ok(mut db) => {
-            match db.add_user(&user.username, &user.password) {
-                Ok(()) => {}
-                Err(e) => return Json(String::from("Username taken")),
-            };
-        }
-        Err(e) => return Json(String::from("Error connecting to database")),
-    }
-    Json(String::from(format!("Created user")))
-}
-
-#[post("/login", format = "application/json", data = "<user>")]
-fn login(user: Json<UserInfo>) -> Json<String> {
-    match database::MyMongo::new() {
-        Ok(mut db) => match db.login(&user.username, &user.password) {
-            Ok(res) => {
-                if res == false {
-                    return Json(String::from("Login failed"));
-                }
-            }
-            Err(e) => return Json(String::from("Login failed")),
-        },
-        Err(e) => return Json(String::from("Error connecting to database")),
-    }
-    Json(String::from(format!("Login success")))
-}
-
-#[post("/update_score", format = "application/json", data = "<score>")]
-fn update_score(score: Json<ScoreUpdate>) -> Json<String> {
-    match database::MyMongo::new() {
-        Ok(mut db) => match db.update_score(&score.username, score.game, score.win) {
-            Ok(res) => {
-                if res {
-                    return Json(String::from("Update success"));
-                } else {
-                    return Json(String::from("Update failed"));
-                }
-            }
-            Err(e) => return Json(String::from("Update failed")),
-        },
-        Err(e) => return Json(String::from("Update failed")),
-    }
-}
-
-#[get("/scores/<username>")]
-fn get_scores(username: &RawStr) -> Json<GameInfo> {
-    let err = GameInfo {
-        username: "".to_string(),
-        xo_wins: -1,
-        xo_loss: -1,
-        xo_ties: -1,
-        to_wins: -1,
-        to_loss: -1,
-        to_ties: -1,
-    };
-    match database::MyMongo::new() {
-        Ok(mut db) => {
-            match db.get_game_score(username.to_string()) {
-                Ok(r) => {
-                    if r.is_none() {
-                        return Json(err);
-                    } else {
-                        return Json(r.unwrap());
-                    }
-                }
-                Err(e) => return Json(err),
-            };
-        }
-        Err(e) => return Json(err),
+        Ok(MyMongo { db: db })
     }
 }
 
@@ -121,11 +44,17 @@ fn main() {
         .allowed_origins(AllowedOrigins::all())
         .allowed_methods(vec![Get, Post].into_iter().map(From::from).collect())
         .allow_credentials(true);
+
     rocket::ignite()
         .attach(cors.to_cors().unwrap())
         .mount(
             "/",
-            routes![index, new_user, login, get_scores, update_score],
+            routes![
+                user_controller::new_user,
+                user_controller::login,
+                game_controller::get_scores,
+                game_controller::update_score
+            ],
         )
         .launch();
 }
